Refactored Sey's Filesize class to fit the project's style-guide.
